fix zh-CN(zh-Hans) and zh-TW(zh-Hant) culture name issue
