+++ test/TestFormats.zh-Hant.cs
@@ -1,35 +1,35 @@
-﻿using Xunit;
+using Xunit;
 
 namespace CronExpressionDescriptor.Test
 {
     /// <summary>
     /// Tests for Chinese Simplified translation
     /// </summary>
-    public class TestFormatsCN : Support.BaseTestFormats
+    public class TestFormatsTW : Support.BaseTestFormats
     {
         protected override string GetLocale()
         {
-            return "zh-CN";
+            return "zh-hant";
         }
 
         [Fact]
         public void TestEveryMinute()
         {
-            Assert.Equal("每分钟", GetDescription("* * * * *"));
+            Assert.Equal("每分鐘", GetDescription("* * * * *"));
         }
 
         [Fact]
         public void TestEvery1Minute()
         {
-            Assert.Equal("每分钟", GetDescription("*/1 * * * *"));
-            Assert.Equal("每分钟", GetDescription("0 0/1 * * * ?"));
+            Assert.Equal("每分鐘", GetDescription("*/1 * * * *"));
+            Assert.Equal("每分鐘", GetDescription("0 0/1 * * * ?"));
         }
 
         [Fact]
         public void TestEveryHour()
         {
-            Assert.Equal("每小时", GetDescription("0 0 * * * ?"));
-            Assert.Equal("每小时", GetDescription("0 0 0/1 * * ?"));
+            Assert.Equal("每小時", GetDescription("0 0 * * * ?"));
+            Assert.Equal("每小時", GetDescription("0 0 0/1 * * ?"));
         }
 
         [Fact]
@@ -53,14 +53,14 @@ namespace CronExpressionDescriptor.Test
         [Fact]
         public void TestEvery5Minutes()
         {
-            Assert.Equal("每 5 分钟", GetDescription("*/5 * * * *"));
-            Assert.Equal("每 10 分钟", GetDescription("0 0/10 * * * ?"));
+            Assert.Equal("每 5 分鐘", GetDescription("*/5 * * * *"));
+            Assert.Equal("每 10 分鐘", GetDescription("0 0/10 * * * ?"));
         }
 
         [Fact]
         public void TestEvery5MinutesOnTheSecond()
         {
-            Assert.Equal("每 5 分钟", GetDescription("0 */5 * * * *"));
+            Assert.Equal("每 5 分鐘", GetDescription("0 */5 * * * *"));
         }
 
         [Fact]
@@ -78,19 +78,19 @@ namespace CronExpressionDescriptor.Test
         [Fact]
         public void TestMinuteSpan()
         {
-            Assert.Equal("在 11:00 AM 和 11:10 AM 之间的每分钟", GetDescription("0-10 11 * * *"));
+            Assert.Equal("在 11:00 AM 和 11:10 AM 之間的每分鐘", GetDescription("0-10 11 * * *"));
         }
 
         [Fact]
         public void TestOneMonthOnly()
         {
-            Assert.Equal("每分钟, 仅在 三月", GetDescription("* * * 3 *"));
+            Assert.Equal("每分鐘, 僅在 3月", GetDescription("* * * 3 *"));
         }
 
         [Fact]
         public void TestTwoMonthsOnly()
         {
-            Assert.Equal("每分钟, 仅在 三月 和 六月", GetDescription("* * * 3,6 *"));
+            Assert.Equal("每分鐘, 僅在 3月 和 6月", GetDescription("* * * 3,6 *"));
         }
 
         [Fact]
@@ -108,103 +108,103 @@ namespace CronExpressionDescriptor.Test
         [Fact]
         public void TestOnceAWeek()
         {
-            Assert.Equal("在 09:46 AM, 仅在 星期一", GetDescription("46 9 * * 1"));
+            Assert.Equal("在 09:46 AM, 僅在 星期一", GetDescription("46 9 * * 1"));
         }
 
         [Fact]
         public void TestDayOfMonth()
         {
-            Assert.Equal("在 12:23 PM, 每月的 15 号", GetDescription("23 12 15 * *"));
+            Assert.Equal("在 12:23 PM, 每月的 15 號", GetDescription("23 12 15 * *"));
         }
 
         [Fact]
         public void TestMonthName()
         {
-            Assert.Equal("在 12:23 PM, 仅在 一月", GetDescription("23 12 * JAN *"));
+            Assert.Equal("在 12:23 PM, 僅在 1月", GetDescription("23 12 * JAN *"));
         }
 
         [Fact]
         public void TestDayOfMonthWithQuestionMark()
         {
-            Assert.Equal("在 12:23 PM, 仅在 一月", GetDescription("23 12 ? JAN *"));
+            Assert.Equal("在 12:23 PM, 僅在 1月", GetDescription("23 12 ? JAN *"));
         }
 
         [Fact]
         public void TestMonthNameRange2()
         {
-            Assert.Equal("在 12:23 PM, 一月 到 二月", GetDescription("23 12 * JAN-FEB *"));
+            Assert.Equal("在 12:23 PM, 1月 到 2月", GetDescription("23 12 * JAN-FEB *"));
         }
 
         [Fact]
         public void TestMonthNameRange3()
         {
-            Assert.Equal("在 12:23 PM, 一月 到 三月", GetDescription("23 12 * JAN-MAR *"));
+            Assert.Equal("在 12:23 PM, 1月 到 3月", GetDescription("23 12 * JAN-MAR *"));
         }
 
         [Fact]
         public void TestDayOfWeekName()
         {
-            Assert.Equal("在 12:23 PM, 仅在 星期日", GetDescription("23 12 * * SUN"));
+            Assert.Equal("在 12:23 PM, 僅在 星期日", GetDescription("23 12 * * SUN"));
         }
 
         [Fact]
         public void TestDayOfWeekRange()
         {
-            Assert.Equal("每 5 分钟, 在 03:00 PM 和 03:59 PM 之间, 星期一 到 星期五", GetDescription("*/5 15 * * MON-FRI"));
+            Assert.Equal("每 5 分鐘, 在 03:00 PM 和 03:59 PM 之間, 星期一 到 星期五", GetDescription("*/5 15 * * MON-FRI"));
         }
 
         [Fact]
         public void TestDayOfWeekOnceInMonth()
         {
-            Assert.Equal("每分钟, 在 第三个星期一 每月", GetDescription("* * * * MON#3"));
+            Assert.Equal("每分鐘, 在 第三個星期一 每月", GetDescription("* * * * MON#3"));
         }
 
         [Fact]
         public void TestLastDayOfTheWeekOfTheMonth()
         {
-            Assert.Equal("每分钟, 每月的最后一个 星期四 ", GetDescription("* * * * 4L"));
+            Assert.Equal("每分鐘, 每月的最後一個 星期四 ", GetDescription("* * * * 4L"));
         }
 
         [Fact]
         public void TestLastDayOfTheMonth()
         {
-            Assert.Equal("每 5 分钟, 每月的最后一天, 仅在 一月", GetDescription("*/5 * L JAN *"));
+            Assert.Equal("每 5 分鐘, 每月的最後一天, 僅在 1月", GetDescription("*/5 * L JAN *"));
         }
 
         [Fact]
         public void TestLastWeekdayOfTheMonth()
         {
-            Assert.Equal("每分钟, 每月的最后一个平日", GetDescription("* * LW * *"));
+            Assert.Equal("每分鐘, 每月的最後一個平日", GetDescription("* * LW * *"));
         }
 
         [Fact]
         public void TestLastWeekdayOfTheMonth2()
         {
-            Assert.Equal("每分钟, 每月的最后一个平日", GetDescription("* * WL * *"));
+            Assert.Equal("每分鐘, 每月的最後一個平日", GetDescription("* * WL * *"));
         }
 
         [Fact]
         public void TestFirstWeekdayOfTheMonth()
         {
-            Assert.Equal("每分钟, 每月的 第一个平日 ", GetDescription("* * 1W * *"));
+            Assert.Equal("每分鐘, 每月的 第一個平日 ", GetDescription("* * 1W * *"));
         }
 
         [Fact]
         public void TestFirstWeekdayOfTheMonth2()
         {
-            Assert.Equal("每分钟, 每月的 第一个平日 ", GetDescription("* * W1 * *"));
+            Assert.Equal("每分鐘, 每月的 第一個平日 ", GetDescription("* * W1 * *"));
         }
 
         [Fact]
         public void TestParticularWeekdayOfTheMonth()
         {
-            Assert.Equal("每分钟, 每月的 最接近 5 号的平日 ", GetDescription("* * 5W * *"));
+            Assert.Equal("每分鐘, 每月的 最接近 5 號的平日 ", GetDescription("* * 5W * *"));
         }
 
         [Fact]
         public void TestParticularWeekdayOfTheMonth2()
         {
-            Assert.Equal("每分钟, 每月的 最接近 5 号的平日 ", GetDescription("* * W5 * *"));
+            Assert.Equal("每分鐘, 每月的 最接近 5 號的平日 ", GetDescription("* * W5 * *"));
         }
 
         [Fact]
@@ -216,37 +216,37 @@ namespace CronExpressionDescriptor.Test
         [Fact]
         public void TestSecondInternvals()
         {
-            Assert.Equal("在每分钟的 5 到 10 秒", GetDescription("5-10 * * * * *"));
+            Assert.Equal("在每分鐘的 5 到 10 秒", GetDescription("5-10 * * * * *"));
         }
 
         [Fact]
         public void TestSecondMinutesHoursIntervals()
         {
-            Assert.Equal("在每分钟的 5 到 10 秒, 在每小时的 30 到 35 分钟, 在 10:00 AM 和 12:59 PM 之间", GetDescription("5-10 30-35 10-12 * * *"));
+            Assert.Equal("在每分鐘的 5 到 10 秒, 在每小時的 30 到 35 分鐘, 在 10:00 AM 和 12:59 PM 之間", GetDescription("5-10 30-35 10-12 * * *"));
         }
 
         [Fact]
         public void TestEvery5MinutesAt30Seconds()
         {
-            Assert.Equal("在每分钟的 30 秒, 每 5 分钟", GetDescription("30 */5 * * * *"));
+            Assert.Equal("在每分鐘的 30 秒, 每 5 分鐘", GetDescription("30 */5 * * * *"));
         }
 
         [Fact]
         public void TestMinutesPastTheHourRange()
         {
-            Assert.Equal("在每小时的 30 分, 在 10:00 AM 和 01:59 PM 之间, 仅在 星期三 和 星期五", GetDescription("0 30 10-13 ? * WED,FRI"));
+            Assert.Equal("在每小時的 30 分, 在 10:00 AM 和 01:59 PM 之間, 僅在 星期三 和 星期五", GetDescription("0 30 10-13 ? * WED,FRI"));
         }
 
         [Fact]
         public void TestSecondsPastTheMinuteInterval()
         {
-            Assert.Equal("在每分钟的 10 秒, 每 5 分钟", GetDescription("10 0/5 * * * ?"));
+            Assert.Equal("在每分鐘的 10 秒, 每 5 分鐘", GetDescription("10 0/5 * * * ?"));
         }
 
         [Fact]
         public void TestBetweenWithInterval()
         {
-            Assert.Equal("每 3 分钟, 在每小时的 2 到 59 分钟, 在 01:00 AM, 09:00 AM, 和 10:00 PM, 在每月的 11 和 26 号之间, 一月 到 六月", GetDescription("2-59/3 1,9,22 11-26 1-6 ?"));
+            Assert.Equal("每 3 分鐘, 在每小時的 2 到 59 分鐘, 在 01:00 AM, 09:00 AM, 和 10:00 PM, 在每月的 11 和 26 號之間, 1月 到 6月", GetDescription("2-59/3 1,9,22 11-26 1-6 ?"));
         }
 
         [Fact]
@@ -258,49 +258,49 @@ namespace CronExpressionDescriptor.Test
         [Fact]
         public void TestMinutesPastTheHour()
         {
-            Assert.Equal("在每小时的 5 分", GetDescription("0 5 0/1 * * ?"));
+            Assert.Equal("在每小時的 5 分", GetDescription("0 5 0/1 * * ?"));
         }
 
         [Fact]
         public void TestOneYearOnlyWithSeconds()
         {
-            Assert.Equal("每秒, 仅在 2013", GetDescription("* * * * * * 2013"));
+            Assert.Equal("每秒, 僅在 2013", GetDescription("* * * * * * 2013"));
         }
 
         [Fact]
         public void TestOneYearOnlyWithoutSeconds()
         {
-            Assert.Equal("每分钟, 仅在 2013", GetDescription("* * * * * 2013"));
+            Assert.Equal("每分鐘, 僅在 2013", GetDescription("* * * * * 2013"));
         }
 
         [Fact]
         public void TestTwoYearsOnly()
         {
-            Assert.Equal("每分钟, 仅在 2013 和 2014", GetDescription("* * * * * 2013,2014"));
+            Assert.Equal("每分鐘, 僅在 2013 和 2014", GetDescription("* * * * * 2013,2014"));
         }
 
         [Fact]
         public void TestYearRange2()
         {
-            Assert.Equal("在 12:23 PM, 一月 到 二月, 2013 到 2014", GetDescription("23 12 * JAN-FEB * 2013-2014"));
+            Assert.Equal("在 12:23 PM, 1月 到 2月, 2013 到 2014", GetDescription("23 12 * JAN-FEB * 2013-2014"));
         }
 
         [Fact]
         public void TestYearRange3()
         {
-            Assert.Equal("在 12:23 PM, 一月 到 三月, 2013 到 2015", GetDescription("23 12 * JAN-MAR * 2013-2015"));
+            Assert.Equal("在 12:23 PM, 1月 到 3月, 2013 到 2015", GetDescription("23 12 * JAN-MAR * 2013-2015"));
         }
 
         [Fact]
         public void TestHourRange()
         {
-            Assert.Equal("每 30 分钟, 在 08:00 AM 和 09:59 AM 之间, 每月的 5 和 20 号", GetDescription("0 0/30 8-9 5,20 * ?"));
+            Assert.Equal("每 30 分鐘, 在 08:00 AM 和 09:59 AM 之間, 每月的 5 和 20 號", GetDescription("0 0/30 8-9 5,20 * ?"));
         }
 
         [Fact]
         public void TestDayOfWeekModifier()
         {
-            Assert.Equal("在 12:23 PM, 在 第二个星期日 每月", GetDescription("23 12 * * SUN#2"));
+            Assert.Equal("在 12:23 PM, 在 第二個星期日 每月", GetDescription("23 12 * * SUN#2"));
         }
 
         [Fact]
@@ -309,56 +309,56 @@ namespace CronExpressionDescriptor.Test
             Options options = new Options();
             options.DayOfWeekStartIndexZero = false;
 
-            Assert.Equal("在 12:23 PM, 在 第二个星期日 每月", GetDescription("23 12 * * 1#2", options));
+            Assert.Equal("在 12:23 PM, 在 第二個星期日 每月", GetDescription("23 12 * * 1#2", options));
         }
 
         [Fact]
         public void TestHourRangeWithEveryPortion()
         {
-            Assert.Equal("在每小时的 25 分, 每 13 小时, 在 07:00 AM 和 07:59 PM 之间", GetDescription("0 25 7-19/13 ? * *"));
+            Assert.Equal("在每小時的 25 分, 每 13 小時, 在 07:00 AM 和 07:59 PM 之間", GetDescription("0 25 7-19/13 ? * *"));
         }
 
         [Fact]
         public void TestHourRangeWithTrailingZeroWithEveryPortion()
         {
-            Assert.Equal("在每小时的 25 分, 每 13 小时, 在 07:00 AM 和 08:59 PM 之间", GetDescription("0 25 7-20/13 ? * *"));
+            Assert.Equal("在每小時的 25 分, 每 13 小時, 在 07:00 AM 和 08:59 PM 之間", GetDescription("0 25 7-20/13 ? * *"));
         }
 
         [Fact]
         public void TestSecondsInternalWithStepValue()
         {
             // GitHub Issue #49: https://github.com/bradymholt/cron-expression-descriptor/issues/49
-            Assert.Equal("每 30 秒, 开始 在每分钟的 5 秒", GetDescription("5/30 * * * * ?"));
+            Assert.Equal("每 30 秒, 開始 在每分鐘的 5 秒", GetDescription("5/30 * * * * ?"));
         }
 
         [Fact]
         public void TestMinutesInternalWithStepValue()
         {
-            Assert.Equal("每 30 分钟, 开始 在每小时的 5 分", GetDescription("0 5/30 * * * ?"));
+            Assert.Equal("每 30 分鐘, 開始 在每小時的 5 分", GetDescription("0 5/30 * * * ?"));
         }
 
         [Fact]
         public void TestHoursInternalWithStepValue()
         {
-            Assert.Equal("每秒, 每 8 小时, 开始 在 05:00 AM", GetDescription(" * * 5/8 * * ?"));
+            Assert.Equal("每秒, 每 8 小時, 開始 在 05:00 AM", GetDescription(" * * 5/8 * * ?"));
         }
 
         [Fact]
         public void TestDayOfMonthInternalWithStepValue()
         {
-            Assert.Equal("在 07:05 AM, 每 3 天, 开始 每月的 2 号", GetDescription("0 5 7 2/3 * ? *"));
+            Assert.Equal("在 07:05 AM, 每 3 天, 開始 每月的 2 號", GetDescription("0 5 7 2/3 * ? *"));
         }
 
         [Fact]
         public void TestMonthInternalWithStepValue()
         {
-            Assert.Equal("在 07:05 AM, 每 2 月, 三月 到 十二月", GetDescription("0 5 7 ? 3/2 ? *"));
+            Assert.Equal("在 07:05 AM, 每 2 月, 3月 到 12月", GetDescription("0 5 7 ? 3/2 ? *"));
         }
 
         [Fact]
         public void TestDayOfWeekInternalWithStepValue()
         {
-            Assert.Equal("在 07:05 AM, 每周的每 3 天, 星期二 到 星期六", GetDescription("0 5 7 ? * 2/3 *"));
+            Assert.Equal("在 07:05 AM, 每週的每 3 天, 星期二 到 星期六", GetDescription("0 5 7 ? * 2/3 *"));
         }
 
         [Fact]
